requirements.txt
aiofile~=0.5.0
aniso860~=7.0.0
async-exit-stac~=1.0.1
async-generato~=1.10
certif~=2020.12.5
charde~=4.0.0
clic~=7.1.2
coloram~=0.4.4
dnspytho~=2.1.0
email-validato~=1.1.2
fastap~=0.65.1
graphen~=2.1.8
graphql-cor~=2.3.2
graphql-rela~=2.0.1
h1~=0.12.0
idn~=2.10
itsdangerou~=1.1.0
Jinja~=2.11.3
MarkupSaf~=2.0.0
orjso~=3.5.2
promis~=2.3
pydanti~=1.8.2
python-doten~=0.17.1
python-multipar~=0.0.5
PyYAM~=5.4.1
request~=2.25.1
R~=1.6.1
si~=1.16.0
starlett~=0.14.2
typing-extension~=3.10.0.0
ujso~=4.0.2
urllib~=1.26.4
uvicor~=0.13.4
watchgo~=0.7
websocket~=8.1
SQLAlchem~=1.4.15